+++ src/experiments/cover_size.cpp
@@ -4,11 +4,11 @@ using namespace std;
 
 void CoverSize::Run(int numberOfInterations, Algorithm &algorithm, Graph &graph)
 {
-  vector<int> coverSizes(numberOfInterations);
+  vector<size_t> coverSizes(numberOfInterations);
   int sum = 0;
   for (int i = 0; i < numberOfInterations; ++i)
   {
-    vector<int> vertices = algorithm.Run(graph);
+    vector<size_t> vertices = algorithm.Run(graph);
     coverSizes[i] = vertices.size();
     sum += vertices.size();
   }
